Profile and plain txt pages added. Links added
Made a few of the pages prettier too. unfortunately with the stick
footer the page has an enormous amount of width. I cant seem to fix it
with out using a fixed to bottom nav bar in its place. It just doesn't
look as nice so I have left it out.